Move item getter to item file
