+++ pages/api/bib/[id]/items.ts
@@ -1,6 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from "next"
 
-import { fetchBibItems } from "../../../../src/server/api/bib"
+import { fetchItems } from "../../../../src/server/api/items"
 
 /**
  * Item fetching route handler for Bib page
@@ -12,7 +12,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   const id = req.query.id as string
   const viewAllItems = req.query?.view_all_items === "true"
 
-  const { status, items } = await fetchBibItems(id, req.query, viewAllItems)
+  const { status, items } = await fetchItems(id, req.query, viewAllItems)
 
   if (req.method === "GET") {
     if (!items?.length || status !== 200) {
